[TASK] Solver Upgrade
 - from now on, SCP solver (http://scpsolver.org/) is used as a Java library wrapper of LP solvers in C language to count the EMD value without need of external call to solve the LP

+++ src/com/kappa_labs/ohunter/server/analyzer/Analyzer.java
@@ -50,6 +50,7 @@ public class Analyzer {
         } catch(Exception e) {
             System.err.println("Could not acquire photos from client: "+e);
             e.printStackTrace();
+            //TODO: lepsi throw nejaka OHException
             return 0;
         }
         
@@ -67,7 +68,7 @@ public class Analyzer {
         
         /* Solve the EMP linear problem and return the final result */
         EMDSolver empm = new EMDSolver(problem);
-        ret = empm.countValue();
+        ret = (float) empm.countValue();
         
         return ret * 10f;
     }
